Add GitHub templates
These GitHub templates aim to improve the level of detail on
pull requests and issues raised by the team, to make the
external community feel more welcome and part of the project.

These are copyright to OpenFaaS Ltd, but we are happy to
contribute them under the license of this project for use by
Packet / Tinkerbell.

This PR also closes a PR raised on the tink repo

Closes: https://github.com/tinkerbell/tink/pull/185

Whilst it is useful to have a global template, individual
templates should still be used in specific repos to gather
diagnostic information and to ask the right questions about
PRs.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>