Split terraform apply in its own workflow
The Terraform workflow was getting a bit hard to code and test with too
many tasks to handle. It had to check for push, pull request, look for
particular labels, send comments and so on.

I decide to split the apply part it its own workflow because it sounds
easier to configure when it has to start, only for a push on the default
branch.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>